Fix datepicker date display on Firefox
Given a .datepicker-input height of 26px, Chrome 41.0.2272.89;
Safari 8.0.3 (10600.3.18); and Firefox 35.0.1 all appear to subtract off
the padding (9 x 2 = 18) and border (1 x 2 = 2) and reach a content
height of 6.

Firefox, however, appears to be the only one of the three that actually
honors it -- the others appear to use a minimum value of 11 or so.

Increase it to 31px so Firefox is left with 11px to honor.

[ci skip]