Test email field's required-ness
Test only whether/not the email field is 'required' (e.g.,
  <input ... required="" .../>
) rather than selenium/Firefox's behavior.

PhantomJS 2.0.0 does not seem to heed the 'required' attribute --
it allows the Submit button to be be pressed & take effect even if the
e-mail field is not a valid e-mail address (e.g., empty or no '@').

Such behavior can be browser-dependent and, as such, deserves
browser-specific testing.