Issue #16, Convert Coffeescript to Javascript
converted all .js.coffeescript files to .js. I did not alter the Gemfile
as I'm unsure if the coffeescript gems may be dependencies or any other
features. Also, I noticed Eco is in the gemlist. If Eco is being used
then there may be coffeescript logic remaining but I was unable to
locate any.