Feature/allow sortable unit activities (#5009)
* Add migration for order_number for unit_activities

So that we can explicitly control sort order

* Set order_number when creating or updating unit_activities

New expectation is that creating or updating units requires a full list of all activities in that unit as part of the POST/PUT call.  These activities are expected to be in an array, and the natural order in that array is used to determine order_number

* Make sure to use index-producing iterators if you want index

We were using iterator functions that didn't provide index values, but then expecting them

* Add unit_activity order_number to teacher lesson query

This both provides the value of, and sorts lessons by, the order_number values recorded in the unit_activity table

* Fix some white space ambiguity

NO CODE CHANGES

* Make activity lessons draggable, and let users save current order

* Add order_number sorting for activities when returning student profile

* Add order_number value to activity return result in student profile

* Add auto-updated structure.sql file

* Add unit_activity order_number to API format test

Make sure that we're checking for this new attribute in API returns.

* Prefer order_number in sorting since it is explicit

And it's also set by the user directly, so we should prefer it pretty highly over other sort params

* Update tests for student profile controller to confirm sort order

These were originally running on just one unit_activity, so the sort logic wasn't being tested at all as far as I can tell.  But it was good logic.  Anyway, I added some additional mocked objects with sorting params attached.

* Add light documentation about order_number in unit_activity

* Prefer unit.created_at over unit_activity.order_number when sorting

This should allow us to keep unit_activities from the same unit grouped together in the sort rather than allowing all order_number==1 unit_activities to sort above all order_number==2 unit_activities, which would disrupt unit groupings.

* Update drag-and-drop to save on drop and provide snackbar feedback

* Update sorting algorithm to only re-sort on drop, not during drag

Sorting during drag results in weird behavior as the underlying array is changing out from under the active drag action

* Update cursor styles for sortable lists

* Minor copy tweak per Tom's request

* Update database migration to be a bit tighter

* Remove extraneous code leftover in the sortable lists

* Update snackbar timeout to match what we're using everywhere else

* Fix student profile sort order to group completed activities

We use max_percentage as a proxy for completion

* Remove JSON parsing since it's handled for us

* Refactor API calls to be a bit cleaner and to use request library

* Update test to match new post body expectation

* Move raw SQL-intensive query from a controller into the model

* Update tests to ensure each sorting criteria is independently tested