Pass specified random seed from Spec to Faker
We have some tests that are failing intermitently because of the data generated in our Factories by Faker, but we can't reproduce the errors to try to fix them.  I've configured Faker to use the same seed as Spec, so that if one is specified it gets passed through.

I've also added an npm command to run tests with a seed so that no one had to look up how to do that in the future.