Update the script to determine which keys still need to be deleted
We were pulling from a static list of keys, but we weren't tracking whether those keys had been deleted or not, so subsequent runs of the script could waste time re-"deleting" keys that were already gone.  The list of keys that SHOULD be deleted is constant, though, so we can pull a list of all keys still live and intersect it with the list of keys that should be deleted to get a list of all keys that STILL need to be deleted.