Remove some extraneous data
This is real data in the real session, but after talking to Emilia and then doing some additional digging it appears that these keys are side effects of some other code and get left on the original data object which is then persisted, so their presence or absence doesn't impact application behavior.  Since the new normalization/denormalization logic does not generate these keys, I'm eliminating them from the test data.