feature/quillconnect/normalize-session-data (#5416)
* Update Connect session to normalize questions

Only save the session ID to the datastore, and dereference those IDs into the expected objects on load.  This code will look for and use old-format sessions if new-format sessions are not found, but will prefer new-format sessions if they exist

* Remove accidentally-committed vim swap file

* Retrieve question data for sessions from redux store

This allows us to avoid doing a second pull of all question data

* Make sure we delete both v2 and v4 sessions when a session delete call occurs

* WIP.  Start working out some tests.

Found the data discrepancy causing the failure.  Now I need to figure out if it matters

* Make the tests work.

Next up, I need to see if the removed keys will impact functionality and if they need to be added to the denormalization script, or if I can just ignore them.

* Update the question loading code to load all question types

Before we were only loading sentence combining questions.  Now we load all types and label them properly.

* Update test to work with new question type logic

* Remove some extraneous data

This is real data in the real session, but after talking to Emilia and then doing some additional digging it appears that these keys are side effects of some other code and get left on the original data object which is then persisted, so their presence or absence doesn't impact application behavior.  Since the new normalization/denormalization logic does not generate these keys, I'm eliminating them from the test data.

* Make some changes for a conservative rollout process

- Don't delete old style sessions anymore, so if we need to rollback our code any users on the new scheme will be able to pick up where they were before the rollout.
- Only use new-style session for a small percentage of session IDs (starting at 1%)

* Add an updatedAt timestamp to new style sessions.

It's something we wish we had on our old sessions, so we might as well start implementing it now.

* Generalize the normalization and denormalization of attempt data

Originally this was special-cased to 'answeredQuestions', but sometimes 'currentQuestion' has attempts, too, so we might as well allow the code to detect their presence and handle them.

* Add test scripts to test real world sessions

We want to know if normalizing and then denormalizing a session leaves it in a state that matches the object in storage.  This script finds out for us.

* Add some additional unit tests

* Tweak the imports in the script to be a bit cleaner

* Add 'esm' package required to run session validation script