Add a migration to establish a unique index on qid+text
This is expected behavior that create_or_increment is trying to execute, but it is subject to race conditions.  To avoid race conditions, we'll need to enforce uniqueness at the datbase level, and in order to add such an index, we also need to resolve any cases extant examples of this constraint being violated.

While we're at it, we should add a validator to the underlying model.