Temporarily attach the location.hash to the API request
We're trying to hunt down the source of these calls to all questions, but don't know which sessions are making that request.  Capturing the location.hash should allow us to figure that out, and Papertrail (our log aggregator) logs the params of all requests, so we just need to send it to dig it out.