Grammar/migrate from firebase (#6061)
* Add some new reducers for more flexible data handling

This lets us add and remove individual questions to the question store

* Add new libs for LMS-based Question API

* Update actions to use LMS API instead of Firebase

* Update components to dispatch all actions

Also make some updates to please the React linter

* Update LMS to properly import Grammar questions from Firebase

Just a tweak to the Firebase import rake task since we need it to take an additional argument now

* Add basic tests for Question actions that go to the API

* Add tests for the QuestionApis

* Rename request mock since it was overriding the request lib

Apparently jest will allow you to create mocks that globally override un-pathed libs (stuff installed via npm).  So if the mock file has the same name as the lib, it'll automatically override, and we don't want that here.