Process prompt_text + entry, but only return errors from entry
In order to validate things like capitalization and non-fragmentary sentences, we need to analyze complete sentences which can be obtained by combining the original prompt text with the student's entry.  However, we don't want to return errors that may be found in the prompt text because the student can't edit that part and could get stuck.