lms/report-downloads-bug (#8783)
* Refactor some of the logic that toggles download links

Pretty sure there's some sort of conditional race condition going on.  On some pages, this always component works exactly as intended, but on some pages it always seems to mis-identify the state that the page is in when it checks to see if it should hide the download links.  The updated code consistently works on all pages.

* Cover condition where class name is passed in to component