Update the way we bubble-up touches with custom callbacks
To avoid some performance issues, we've had to write some custom callbacks to simulate `touch` behavior on a couple of models.  One of those custom callbacks wasn't being called in all situations.  The hope is that this fixes our cache invalidation process so that we can turn caching back on if this code doesn't cause performance issues.