Check to make sure that underlying classrooms exist before touching
Classroom has a `cascade: :destroy` chain that, among other things, destroys related classroom units.  A destruction event causes a commit, which triggers our after_commit touch function.  However, in this scenario, the classroom that we specify to be touched has been destroyed, so obviously can't be updated.  This guard clause make sure we don't cause an error in these cases.