lms/confirm-touch-chains (#8832)
* Update the way we bubble-up touches with custom callbacks

To avoid some performance issues, we've had to write some custom callbacks to simulate `touch` behavior on a couple of models.  One of those custom callbacks wasn't being called in all situations.  The hope is that this fixes our cache invalidation process so that we can turn caching back on if this code doesn't cause performance issues.

* Revert a change to this test that was made to handle an after_touch

Now that we've tweaked the after_touch not to trigger after_save hooks, this change is no longer needed and we can use the original version of this test

* Consolidte after_touch + after_save hooks into a single after_commit

* Check to make sure that underlying classrooms exist before touching

Classroom has a `cascade: :destroy` chain that, among other things, destroys related classroom units.  A destruction event causes a commit, which triggers our after_commit touch function.  However, in this scenario, the classroom that we specify to be touched has been destroyed, so obviously can't be updated.  This guard clause make sure we don't cause an error in these cases.