evidence/fix-plagiarism-with-repeated-words (#9034)
* Update plagiarism matching so that we count repeated words

We use the length of the intersection of two arrays to determine if "enough" words are the same between those arrays to be worth doing a Levenshtein distance calculation on them.  Naive array intersections in Ruby (using &) collapse duplicate values, so [1,1] & [1,1] becomes [1].  This leads to arrays that are shorter than we expect, and thus caused the algorithm to decide that it wasn't worth doing a Levenshtein even when we wanted it to.  This set of changes re-hydrates the intersection value with duplicates so that the length is what we would expect.

* Fix linting error

* Refactor based on a suggestion from Dan

* More tweaks based on Dan feedback, fix some tests

* Fix linting error

* Small optimization tweak from Dan