Add a call to save ActiveActivitySession when nextStep is set
Without this call, a user who completes a prompt, clicks "next", but then leaves the page or clicks "save and exit" will not have the start of the new prompt saved and will resume the activity with the final feedback of the prompt they completed.  Adding this here means that if you click "next" to start a new prompt, the activeStep gets updated in the saved session even if you take no further action.