lms/show-evidence-completed-in-activity-feed (#9579)
* Ensure that Evidence activities show as 'completed' in activity feed

There was an early return to protect this code from errors with calculating stuff around percentages, but it was accidentally early-returning on Evidence activities (which we explicitly don't record a `percentage` value for).  This wasn't caught in existing specs because the factory to create Evidence activity sessions wasn't explicitly using `nil` for percentage.  Updated the factory to confirm existing specs would catch this condition, then fixed the bug

* Fix a minor typo