Lms/expand district nces id size (#9598)
* Add a migration to make districts.nces_id a bigint instead of int

We have a known school that is paying for premium that has a 12-digit NCES ID, and that ID can't actually be set on the district without this migration.  Having an NCES ID mismatch is disrupting our Clever integration

* Commit documentation updates from an old migration

* Fix linting error

* Update migration to no-op when rolled back

Because once data of the new size gets into the database, the rollback won't be possible until/unless it's removed.