Update Google auth code to handle `sales-contact` users
Basically, if you try to "sign in" and have a sales-user record, treat it as if you don't have a user at all (and show an error).  If you "sign up" and you have a sales-user record, update the `role` on that record to the appropriate type based on sign-up.