Treat non-authenticating user roles as non-existant during sign-in
We want a user who signs in with email/password for an email that is assigned to a `sales-concat` user type to get the same experience as someone signing in with an email that doesn't correspond to any user.  This is because we want to direct these users to the sign-up flow if they actually want to use Quill.