Lms/enable new sales contacts (#9621)
* Re-enable the new sales contact form and Intercom webhooks

* WIP attempts to get tests for new Google expectations to work

* Update Google auth code to handle `sales-contact` users

Basically, if you try to "sign in" and have a sales-user record, treat it as if you don't have a user at all (and show an error).  If you "sign up" and you have a sales-user record, update the `role` on that record to the appropriate type based on sign-up.

* Update password reset to treat non_authenticating Users as missing

We have a specific message that we use when a user tries to re-set a password for a "non-existing" user, and we want to use that same message if the user has a non-authenticating user type like "sales-contact" since they don't know they have an account and we want them to go through the sign-up process.

* Treat non-authenticating user roles as non-existant during sign-in

We want a user who signs in with email/password for an email that is assigned to a `sales-concat` user type to get the same experience as someone signing in with an email that doesn't correspond to any user.  This is because we want to direct these users to the sign-up flow if they actually want to use Quill.

* Fix linting errors

* Simplify some early return logic to avoid linting errors

* Update code based on review feedback

* After discussion and testing with Peter S, disable this link again

We'll turn it back on once we figure out with Vitally what we actually want to do around managing the receipt of these forms