Make sure to permit expected params for handling Lessons concept_results
The function that this param key was being passed into before was doing a synchronous `create` call for models.  Apparently you can pass an ActionController::Parameters object to ActiveRecord::Model.create and it gets interpreted just fine.  But now we handle the creation in Sidekiq workers, and serializing a ActionController::Parameters that have nested hashes apparently calls `.to_s` on those hashes, resulting in a not-easy-to-de-serialize argument on the Sidekiq queue.  We need to properly ensure that the ActionController::Parameters version of the data gets converted into a simple Hash before we serialize the data onto the Sidekiq queue.