Allow ConceptResults to init when metadata is incorrectly serialized
An earlier bug caused the metadata key to serialize it's data with `.to_s` instead of `.to_json`, and queue items with these malformed payloads into the Sidekiq queue.  We want to actually write that data to the database properly, so we need special case handling to do so.  The bug that causes this issue should also be resolved when we deploy this code, so it will only be temporarily required