Lms/fix concept result handling for lessons (#9679)
* Make sure to permit expected params for handling Lessons concept_results

The function that this param key was being passed into before was doing a synchronous `create` call for models.  Apparently you can pass an ActionController::Parameters object to ActiveRecord::Model.create and it gets interpreted just fine.  But now we handle the creation in Sidekiq workers, and serializing a ActionController::Parameters that have nested hashes apparently calls `.to_s` on those hashes, resulting in a not-easy-to-de-serialize argument on the Sidekiq queue.  We need to properly ensure that the ActionController::Parameters version of the data gets converted into a simple Hash before we serialize the data onto the Sidekiq queue.

* Allow ConceptResults to init when metadata is incorrectly serialized

An earlier bug caused the metadata key to serialize it's data with `.to_s` instead of `.to_json`, and queue items with these malformed payloads into the Sidekiq queue.  We want to actually write that data to the database properly, so we need special case handling to do so.  The bug that causes this issue should also be resolved when we deploy this code, so it will only be temporarily required

* Remove a local variable that we can just directly reference

* Add a bit more to the Lesson concept_result spec so it's more robust

* Refactor our worker to avoid a deprecation warning

* Fix linting error

* Fix linting error