Use only visible ActivitySessions for Lesson student rosters
We use ActivitySession UIDs to validate that a given student is supposed to join a given Lesson, but we ran into a bug where a student had a deleted (visible == false) ActivitySession in the db that was getting used instead of the active session.  This meant that the student was being told that they weren't supposed to join the Lesson because of the mismatch in uids.