Lms/pull stripe invoices (#10213)
* Update backend for Subscription creation for Stripe Invoice fetching

Add code to fetch total payment information and purchaser email from Stripe when manually creating a subscription based on a Stripe Invoice.  Provide some validation of Stripe Invoice ID (it has to point to a real Stripe Invoice), and error if it doesn't.  This will require some changes to the admin tools to handle that error when it comes up.

* WIP modify UI for creating new invoices by staff in LMS

* Update Subscription creation to allow for invalid stripe_invoice_ids

Per updates to the spec

* Update staff tools for creating Subscriptions

We want to auto-pull data from Stripe if a valid Stripe Invoice ID is provided

* Update the correct Subscription controller for our fields

* Revert changes to irrelevant controller for this code

* Add onChange handlers to new input fields

So that the submitted payload is updated with the values from those fields

* Add a bunc of temporarl logging to debug in staging

* Revert "Add a bunc of temporarl logging to debug in staging"

This reverts commit 5c21fdad311c6474828200a2b10a0c08614f8fcb.

* Remove unique index on Subscription.stripe_invoice_id

Now that we're going to begin actually recording these for some subscriptions, we don't want to restrict what Ops and Partnerships can put into these fields

* Add Subscription.purchase_order_number data to Vitally sync payloads

Since we've added that column in this PR and want that data to sync to Vitally when it exists

* Be more selective in which error classes we swallow

* Fix linting errors

* Update db/structure.sql from `develop` to fix discrepancy with local

* Modify guard clause to work with new definition of stripe?

* Clean up some specs

* Change the redirect path after creating subscriptions

So that instead of going to the edit page for the subscription, it goes back to the detail page for either the School or District the subscription was created for.  This is a request from product management not originally included in the spec.

* Remove no-longer-accurate copy at the request of Product