Tweak the latency status controller to 400 in certain conditions
As far as we can tell, the easiest way for us to construct a synthetic alarm in NewRelic that cares about multiple data points is to do so using NRQL, but the features that we can query in NRQL do not include the response body of a synthetic check.  We need to differentiate the response status code in "good" situations vs "bad" ones to work up an NRQL query that can figure out if we've had multiple failures in a row, which is what we want to be querying.