Update the Clever TeacherUpdater so that it doesn't make admins teachers
This code was overly-aggressive about assigning the "teacher" role to users when they logged in.  This change makes it so that they only get made a teacher if they're not already some other kind of "school staff" type of user (as proxied by having a role in TEACHER_INFO_ROLES).