Tweak ClassroomManager teacher info modal to appearance logic
We used to be able to depend on the simple presence/absence of teacher_info records for the teacher to determine whether the teacher should see the modal asking for their grade levels and subject areas.  Now that we're storing teacher_info for additional info not related to this modal (email frequency), we need a slightly more precise thing to check to figure out whether the user needs to see the modal.  I've settled on minimum_grade_level for now.