Lms/add teacher notification email frequency (#10525)
* Add a new TeacherInfo field for storing email notification frequency

Also update the TeacherInfosController `create` action so that it won't create additional records if one already exists for a given user.

* Set up code to set default values for notification_frequency on new users

* Remove two migrations from structure from a different branch

* Remove more migrations from other branches from structure.sql

* Fix db/structure, hopefully

* Clean up specs to work with new more common TeacherInfo records

* Tweak ClassroomManager teacher info modal to appearance logic

We used to be able to depend on the simple presence/absence of teacher_info records for the teacher to determine whether the teacher should see the modal asking for their grade levels and subject areas.  Now that we're storing teacher_info for additional info not related to this modal (email frequency), we need a slightly more precise thing to check to figure out whether the user needs to see the modal.  I've settled on minimum_grade_level for now.

* Update more specs using teacher_info

* Make tweaks based on code review

* Use update action for teacher info when signing up

Since we're already creating a teacher_info record for new users in a hook, so making a create call here will error