Lms/admin snapshot report backend milestone2 redux (#10626)
* Get basic queries committed

* Add a utitlity for generating consistent Snapshot cache keys

* Add a worker to generate and cache Count snapshots

Then send a Pusher notification to let the front-end know that it should re-request the payload

* Add a controller to make Snapshot data requests to

It checks the cache to see if the data is available, and if the data is missing it kicks off a Sidekiq worker to calculate said data

* Update cache key generation to allow for custom timeframes

* Remove old spec file for query we're moving into a later milestone

* Fix linting

* Make sure timeframes end at the end of previous day

Instead of at time of query run

* Make sure to generate cache key in one place

Generate it in the controller and pass it to the worker to guarantee that the keys match

* Require school_ids list as part of queries

* Tweak cache key generation to be more precise

Focus on timeframes and school lists instead of users.  Note that this requires explicit school_ids to be passed in from the front-end for these reports, but that was already the plan

* Add authentication and validation to the controller

* Revise the params we pass into the worker

* Revert "Revise the params we pass into the worker"

This reverts commit 752af7bd7fd508efddab446477246028423017c2.

* Allow a nil previous-timeframe to be passed to the cache worker

And in cases when it is, only calculate current timeframe and not previous.  This allows us to execute queries only a current timeframe desired

* Include previous timeframe start in cache keys

This accounts for the possibility of ambiguous previous timeframes when dealing with custom entries

* Move VCR config changes into the rails_helper

This vastly simplifies ourr external_api specs

* Update spec expectations to match current data responses

If this drifts again we'll need to do a deeper dig to see why results are changing over time

* Refactor snapshot queries to not use user_id

Just rely on a list of school_ids to calculate values.  We'll rely on the controller to authorize a given current_user to get access to a given set of schools in a report

* Fix linting errors

* Move magic query inputs into a shared context

* Add an comment explaining the source of the values in shared context

* Updates based on PR feedback

* Minor tweak to timeframe logic for "this X" items

Make the previous timeframe period the same length as the current one so that we are comparing similar lengths of time against each other

* Move Timeframe logic into its own lib module

* Tweak worker for payload shape

* WIP queries for Milestone 2

* Make some small tweaks to the queries

* Add specs for new TopX queries

* Refactor controller to re-use code between actions

* Add worker and spec for TopX cache worker

* Add specs for the TopX controller action

* WIP adding spec to TopX queries

* Add specs to cover top_x queries using new CTE structure

* Move specs from `external_api` tag to `big_query_snapshot`

* Make a couple of additional tweaks based on PR feedback

* Fix linting errors

* Fix another linting error

* Make some tweaks to specs based on PR feedback

* More PR feedback

* Make too_old and too_new values part of a context

So that we can use let in the context instead of = inside a specific case

* Refactor the specs to use more shared context within each spec file

* Define our block properly

* Rely on shared context to define runner and results

Also do some cleanup in the Snapshot Options shared context so that all items instantiate as expected.  Also, also rename "params" context variable to "query_args" since that's more descriptive.