Lms/add admin snapshot filter options (#10679)
* Refactor OptionsQuery to take filter values

* Add new options queries

* Update SnapshotsController#options action to take filter params

This allows the back-end to update the front-end on which higher-specificity filters should be shown to the user based on broader filters that have already been applied.

* Appease Rubocop linter.

The current configuration we have doesn't allow for open-ended ranges

* Some more linter clean-up

The big one here is to declare our test class in a `let`

* Use named params for queries since the list of args is getting long

* WIP tweak options controller to use PostgreSQL rather than BigQuery

It's less brittle and also more up-to-date, which allows a newly-created admin to immediately access their reports

* First pass at SnapshotController#options action in PostgreSQL

* Refactor SnapshotController#options to use PostgreSQL

* Remove BigQuery implementation of options queries

It makes more sense to pull them out of PostgreSQL for both performance and usability reasons

* Add additional filters to PeriodQuery

To allow for filtering by classroom_id or teacher_id.  Also add spec coverage for root PeriodQuery behavior

* Clean up trailing whitespace

* Clean up specs to work with the updated param passing

* Update cache logic to handle new filter params

* Update specs for cache key generation based on new params

* Refactor cache lib to meet linter max arg limits

* Fix linting error, add a spec for cache keys with no extra filters

* Re-add an accidentally removed grade level

* Update CacheKeys lib to DRY it out

* Refactor PeriodQuery for cleaner syntax around early returns

Use .blank? over .nil? || .empty?

* Refactor workers to have a similar signature to cache key generation

Specifically, pull `school_ids` out of the `filters` because school_ids is a required param for all queries.

* Refactor controller to simplify code by breaking it into more methods