Lms/admin snapshots filters teachers without classrooms (#10988)
* Include teachers without classrooms in filters

* Update logic for when "all" items are selected from a filter

Specifically, we allow "selected" items to be a super-set of "all" items in cases where selecting (say) a single school reduces the number of teachers being shown as "all" because we only want to show the teachers from that school.  This code recognizes that "all" are selected if every item in the "all" is included in "selected" even if "selected" has additional items

* Fix linting error

* Tweak timeframe calculation so that custom values go to day boundary

We were just taking raw values from the datepicker, but in order to be consistent with other values, we want to set bounds to start and end of day.  This means that if you do a custom timeframe that corresponds (date-wise) with with a defined timeframe, you should get the same results

* Name variable something more useful than `i`