Add a max-width to team-member divs
We were relying on image size of 219x219px to set the size of these elements, but we're considering using 2x images for better display on hi-res displays where normal images can look fuzzy.  In order to support that, we'll need the CSS of the page itself to control the size of these divs.