Lms/split snapshot count timeframes (#11138)
* Refactor AdminSnapshot so that count and previous_count are separate

This will make it easier to request those two data points in parallel rather than serially from the front-end, which should lead to better performance

* Update controller spec for tweaked signatures and names

* Return early if timeframes are nil

This is only expected to happen in cases where the selected timeframe is "all-time" in which case there is no timeframe for previous_count

* Update front-end to request current and previous period data separately

* Add some additional params to component for testing

* Update change class logic

This is really an update for testing.  In the old system, changeDirection was nulled by default and wasn't set during some tests.  In the new system, effects set the value of changeDirection all the time, so keeping consistent class behavior required a tweak

* Update test snapshots with new parameters

* Update method signature per linting rule Style/OptionalBooleanParameter

* Tweak spec to align with new payload shape expectations

* Updates based on PR feedback

Specifically, fold current and previous timeframe controller actions and workers into singles one with params to differentiate them

* Add unused method param to allow for matching signatures

* A bit more updating based on code review

* Fix a small bug in async direction calculation

* Fix linting error

* Fix a spec to work with the updated timeframe lib code

* Fix signature mis-match in passing previous_timeframe param