Refactor Snapshot Pusher events to include query
So that we can reduce how many triggers a given widget has to respond to.  We were letting all Count widgets get the Pusher "count cache finished" event, and then checked the payload in the Pusher message to see if the query in question was relevant to the widget.  This way individual widgets just ignore Pusher messages for other widgets of the same type.