Lms/fix admin diagnostic percentage bug (#11209)
* Add timeframe start and end info to snapshot pusher payloads

We were differntiating different timeframes by name, but that was making it impossible to differentiate between different configurations of "custom".  This was causing all previously-run custom timeframes from a given session to trigger their callbacks for any new custom timeframe Pusher payload.  Including start and end in the Pusher payload lets us further restrict callbacks

* Refactor Snapshot Pusher events to include query

So that we can reduce how many triggers a given widget has to respond to.  We were letting all Count widgets get the Pusher "count cache finished" event, and then checked the payload in the Pusher message to see if the query in question was relevant to the widget.  This way individual widgets just ignore Pusher messages for other widgets of the same type.

* Add hash check logic to Data Export worker

So that we aren't sending giant filter payloads in that worker

* Supplement hashed Pusher payloads with custom timeframe params

Instead of the timeframe for the calculation, which "previous" requests don't have context for (when you pick a custom timeframe on the front-end, the front-end doesn't know what the "previous" timeframe will look like) use the actual "custom timeframe" value passed from the front-end

* Move custom_timeframe Pusher data into the hash

Rather than having it be a separate part of the payload

* Fix linting errors

* Compare custom timeframe strings directly

Instead of trying to sync up complicated string manipulation between front and back ends

* Remove leftover dead function

* Fix missing , in array

* Remove unused code