Modify "completion" logic to include only currently-assigned students
The old logic counted a completion even if the user was no longer assigned to a given activity (such as the student leaving a classroom).  This was resulting in counts of completion higher than assignment in some cases, and also violates how things work for teacher-level reports (which exclude no-longer-assigned students).  So this change resolves that issue.